+++ src/Exception/ContractsExceptionInterface.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace ComposerUnused\Contracts\Exception;
+
+use Throwable;
+
+interface ContractsExceptionInterface extends Throwable
+{
+}