+++ src/LinkInterface.php
@@ -7,4 +7,9 @@ namespace ComposerUnused\Contracts;
 interface LinkInterface
 {
     public function getTarget(): string;
+
+    /**
+     * Returns the line number of the target from composer.json
+     */
+    public function getLineNumber(): int;
 }