+++ CHANGELOG.md
@@ -2,6 +2,20 @@
 
 ## [Unreleased] - TBA
 ### Added
+### Changed
+### Removed
+
+## [0.2.0] - TBA
+### Added
+- Added `LinkInterface::getLineNumber(): int`
+- Added `PackageInterface::getRequire(string $name): LinkInterface`
+- Added `ContractsExceptionInterface`
+- Added `LinkNotFoundException`
+### Changed
+### Removed
+
+## [0.1.0]
+### Added
 - Added `LinkInterface`
 - Added `PackageInterface`
 - Added `RepositoryInterface`